+++ src/lib.rs
@@ -2,6 +2,11 @@
 
 # Crate features
 
+* **propagate** -
+  Enables the distributed tracing context propagation, linking spans across service boundaries.
+  When enabled the trace context is injected into outgoing requests, and extracted from incoming
+  requests. Enabled by default.
+
 * **axum** -
   Enables the [`axum`] integration. Trace and metrics will contain the
   `http.route` attribute, populated with the path in the router that matches