+++ src/trace/propagate.rs
@@ -1,5 +1,3 @@
-//! Implementation of fields injector.
-
 use std::str::FromStr;
 
 use http::{HeaderMap, HeaderName, HeaderValue};
@@ -13,5 +11,22 @@ impl opentelemetry::propagation::Injector for HeaderInjector<'_> {
                 self.0.insert(header_name, header_value);
             }
         }
+    }
+}
+
+pub struct HeaderExtractor<'a>(pub &'a HeaderMap);
+
+impl opentelemetry::propagation::Extractor for HeaderExtractor<'_> {
+    fn get(&self, key: &str) -> Option<&str> {
+        self.0
+            .get(key)
+            .and_then(|header_value| header_value.to_str().ok())
+    }
+
+    fn keys(&self) -> Vec<&str> {
+        self.0
+            .keys()
+            .map(|header_name| header_name.as_str())
+            .collect()
     }
 }