+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Update [`opentelemetry`]` to v0.32.0
 - Update [`tracing-opentelemetry`]` to v0.33.0
 - Fix features `Cargo.toml`
+- Make the context propagation optional, enabled by default.
 
 ## 0.9.0